Fixes Scanner bug relating to C++ case() syntax
In many languages, case(a,b,c): means "if it is a or b or c." In C++, for some reason, it evaluates equivalently to case(c): . My compiler failed to even report that this was a problem; I only found it during runtime debugging. What the hell?

Fixing this meant using regexes. I'll make a generic one and stick it in a comment later for when we may need to do this again for UTF-8 compliance.