Fixes tokenheader misalignment in Parser
This was caused by the new BinExp stuff;

In general, the fix meant converting most read() functions into taking a $here and $there to indicate the range of tokens it should parse for its thing within. After the read() finishes is when it actually updates tokenheader to point towards the thing immediately after read()'s thing it read in. This is a standard I'm going to enforce and continue using as the Parser grows in complexity.