Fully implements Handle, CallExpression, MemberAccess, etc.
Getting the location of things and then accessing them has never been easier!

This gets the Interpreter and AST ready to start running object-oriented code; or at least, code that has more than /main() in it.

Does not seem to break any other part of João just yet, so, that's good!