BIG: Makes Object inheritence work as intended
Making the algorithm within ObjectTree took me a lot of thought, but now that I've written it I'm confident that I've made something that is both very terse, very fast, and very powerful.

However, affecting something as deep to the language as its OOP inheritence scheme meant touching many files. Not sure yet if this is buggy or not.