First attempt at a BIG rewrite of the Handle system
I've come to the realization that the correct way of handling all this Handle nonsense is to just have handle() functions return Value&, since that actually satisfies everything you'd want to do with a handle.

This required giving Value the ability to store a Function pointer in a super secret capacity. Not really working so far but I wanted to have a savestate here for this *massive* change to reference resolution.