Long comments permit whitespace before their invokation
Previously, if the line that a long comment began on did not start with precisely the characters "###", it would throw an error about how long comments cannot occur on a line containing functional code, even if it's simply padded with whitespace characters.

Now, just having some tab characters at the front no longer throws this error. It is still the case that functional code cannot co-exist on the same line as a long comment.