Gets a basic GC system going for Object and String
This system, rather than having a proper GC class & object, implicitly resolves the collection through pointer caches & reference counters maintained by the Value class, based on construction, destruction, copying, and assignment.

For some reason I was lead to believe that this is was the easiest way of implementing this.

Seems to mostly work, but there are apparent bugs in how /error objects are over-destructed (perhaps due to them being made natively?)