Adds some novel documentation
Been trying to write these docs for a while but gave up on it in part because I didn't really like how the HTML generators were handling things.

I've finally decided to say screw it and write it all manually in HTML-CSS-JS. So far, I'm pretty satisfied with the results.