Sweeps up all memory leaks
It's official -- Joao leaves absolutely nothing for __exit to clean up, not a singular byte.

This is not really important to the interpreter -- in fact it might just make it slower -- but it does make it more possible for Joao to run as a service or daemon of some sort. Exciting!