Refactors unit tests, adds fallible ones
Meson purposefully makes it difficult to access all files of the same ending in a directory, which makes programmatically generating a list of tests rather difficult.

Regardless, this new setup should make it easier to do that later on, and the meson file now looks a fair bit more legible :^)

Also added a new must-fail test.